[libyafuse2]: add support for 'statfs' and 'stat'.

+++ libyafuse2/libext4/debug.c
@@ -560,7 +560,7 @@ void ext4_show_inode_stat(struct ext4_super_block *sb, uint64_t ino, struct ext4
   const char *str = NULL;
   time_t tm = 0;
 
-  fprintf(stdout, "Inode %5lu: ", ino);
+  fprintf(stdout, "Inode %5llu: ", (long long unsigned)ino);
 
   str = NULL;
   fprintf(stdout, "type: ");
